refactor(core): add submit-error code, hoist the error normalizer
Add AttaformErrorCode.SubmitError ('atta:submit-error') for the
upcoming throw-surfacing path, and hoist normalizeErrorInput /
normalizeErrorInputs out of build-form-api's closure into errors.ts,
beside toError. Both now take formKey and a defaultCode as required
params; setErrors rewires to the shared version with UserError. No
behavior change.
